Renamed Commands to <action><Subsystem> format and Renamed loader -> shooter and Shooter -> Flywheel
