+++ Warbots2020/src/main/java/frc/robot/commands/AutonomousCommand.java
@@ -5,51 +5,31 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.intake;
+package frc.robot.commands;
 
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Intake;
+import frc.robot.commands.drivetrain.DriveCommand;
+import frc.robot.subsystems.DriveTrain;
 
-public class Capture extends CommandBase
-{  
+public class AutonomousCommand extends DriveCommand 
+{
   //region Constructors
-  public Capture(Intake i) 
+  public AutonomousCommand(DriveTrain dt) 
   {
-    intake = i;
-    addRequirements(intake);
+    super(dt);
   }
   //endregion
 
   //region Overrides
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() 
-  {
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() 
   {
-    intake.intake();;
+    driveTrain.arcadeInput(.5, 0); //TODO Set default autonomous parameters in constants.java
   }
 
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) 
-  {
-    intake.end();
-  }
-
-  // Returns true when the command should end.
   @Override
   public boolean isFinished() 
   {
     return false;
   }
   //endregion
-
-  //region Fields
-  private final Intake intake;
-  //endregion
-}   
+}
