+++ Warbots2020/src/main/java/frc/robot/commands/CaptureIntake.java
@@ -5,56 +5,51 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.shooter;
+package frc.robot.commands;
 
-import java.time.LocalDateTime;
+import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.Intake;
 
-import frc.robot.subsystems.Shooter;
-
-public class SpinUp extends ShooterCommand
-{
-  public LocalDateTime endTime;
+public class CaptureIntake extends CommandBase
+{  
   //region Constructors
-  public SpinUp(Shooter s, double speed) 
+  public CaptureIntake(Intake i) 
   {
-    super(s, speed);
+    intake = i;
+    addRequirements(intake);
   }
   //endregion
 
   //region Overrides
   // Called when the command is initially scheduled.
   @Override
-  public void initialize()
+  public void initialize() 
   {
-    resetEndTime();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute()
+  public void execute() 
   {
-    shooter.setShootSpeed(targetVelocity);
+    intake.intake();;
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) 
   {
-    shooter.setShootSpeed(0);
+    intake.end();
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() 
   {
-    return LocalDateTime.now().isAfter(endTime);
+    return false;
   }
   //endregion
 
-  //region Methods
-  public void resetEndTime()
-  {
-    endTime = LocalDateTime.now().plusSeconds(10);
-  }
+  //region Fields
+  private final Intake intake;
   //endregion
 }
