+++ Warbots2020/src/main/java/frc/robot/commands/ReleaseLowerArmClimber.java
@@ -5,52 +5,47 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.climber;
+package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.Climber;
 
-public class Retract extends CommandBase
-{
-  //region Constructors
-  public Retract(Climber c, double speed) 
+public class ReleaseLowerArmClimber extends CommandBase {
+  /**
+   * Creates a new ReleaseLowerArm.
+   */
+  public ReleaseLowerArmClimber(Climber c)
   {
     climber = c;
     addRequirements(climber);
-    targetVelocity = speed;
   }
-  //endregion
 
-  //region Overrides
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() 
+  public void initialize()
   {
+
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() 
+  public void execute()
   {
-    climber.setSpeed(targetVelocity, 10000);
+    climber.setAngleLower(90);
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) 
+  public void end(boolean interrupted)
   {
+
   }
 
   // Returns true when the command should end.
   @Override
-  public boolean isFinished() 
+  public boolean isFinished()
   {
-    return climber.atSetPosition();
+    return climber.atSetPointLower();
   }
-  //endregion
-
-  //region Fields
-  private final Climber climber;
-  private double targetVelocity;
-  //endregion
-}   
+  public final Climber climber;
+}
