+++ Warbots2020/src/main/java/frc/robot/commands/StuffFlyWheel.java
@@ -5,49 +5,49 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.intake;
+package frc.robot.commands;
+
+import java.time.LocalDateTime;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Intake;
+import frc.robot.robot.Constants;
+import frc.robot.subsystems.FlyWheel;
 
-public class Eject extends CommandBase
+public class StuffFlyWheel extends CommandBase 
 {
-  //region Constructors
-  public Eject(Intake i) 
+  FlyWheel shoot;
+  LocalDateTime endTime;
+  public StuffFlyWheel(FlyWheel s) 
   {
-    intake = i;
-    addRequirements(intake);
+    shoot = s;
+    addRequirements(s);
   }
-  //endregion
 
-  //region Overrides
   // Called when the command is initially scheduled.
   @Override
   public void initialize() 
   {
+    //endTime = LocalDateTime.now().plusSeconds(5);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() 
   {
+    shoot.setShootSpeed(Constants.ShooterConstants.stuffSpeed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) 
   {
+    shoot.setShootSpeed(0);
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() 
   {
-    return false;
+    return false;//LocalDateTime.now().isAfter(endTime);
   }
-  //endregion
-
-  //region Fields
-  private final Intake intake;
-  //endregion
-}   
+}
